+++ src/main/java/racingcar/domain/Race.java
@@ -1,4 +1,4 @@
-package racingcar.model;
+package racingcar.domain;
 
 import java.util.Comparator;
 import java.util.List;
