+++ src/test/java/racingcar/view/InputProcessTest.java
@@ -3,7 +3,7 @@ package racingcar.view;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import racingcar.model.Car;
+import racingcar.domain.Car;
 
 import java.io.ByteArrayInputStream;
 import java.io.ByteArrayOutputStream;
